test: extract test_rc_returns_packed_buffer to unit/ragged/
